radar-main/pyproject.toml
[project]
name = "trigger-radar"
version = "0.8.4"
description = "Industry-first trigger radar for CAR-T and T-cell engagers"
requires-python = ">=3.10"
dependencies = ["requests>=2.31.0","python-dateutil>=2.9.0.post0","PyYAML>=6.0.1"]

[tool.setuptools.packages.find]
where = ["."]
